Move populate method to MethodWidget
